Move data retrieval fns to a new api module.
