--- static/js/ui/comp/index.js
@@ -1,7 +1,0 @@
-import AppAutocomplete from "./AppAutocomplete";
-import SearchForm from "./SearchForm";
-
-export default {
-  AppAutocomplete,
-  SearchForm,
-};
