--- static/js/ui/index.js
@@ -1,9 +1,0 @@
-import a11y from "./a11y";
-import comp from "./comp";
-import conditional from "./conditional";
-
-export default {
-  a11y,
-  comp,
-  conditional,
-};
